Changed defense type to int

+++ Hero.cpp
@@ -2,8 +2,8 @@
 #include "JSON.h"
 #include <iostream>
 
-Hero::Hero(const std::string &name, const int health, const int damage, const double atkcooldown, const double defense,
-           const int ExperiencePerLevel, const int HealthPointBonusPerLevel, const int DamageBonusPerLevel, const double ColdownMultiplierPerLevel, const double DefenseBonusPerLevel)
+Hero::Hero(const std::string &name, const int health, const int damage, const double atkcooldown, const int defense,
+           const int ExperiencePerLevel, const int HealthPointBonusPerLevel, const int DamageBonusPerLevel, const double ColdownMultiplierPerLevel, const int DefenseBonusPerLevel)
     : Monster{name, health, damage, atkcooldown, defense}, maxHealth(health), ExperiencePerLevel(ExperiencePerLevel), HealthPointBonusPerLevel(HealthPointBonusPerLevel),
       DamageBonusPerLevel(DamageBonusPerLevel), ColdownMultiplierPerLevel(ColdownMultiplierPerLevel), DefenseBonusPerLevel(DefenseBonusPerLevel)
 {
@@ -58,10 +58,10 @@ Hero Hero::parse(const std::string &fileName)
         data.get<int>("base_health_points"),
         data.get<int>("base_damage"),
         data.get<double>("base_attack_cooldown"),
-        data.get<double>("defense"),
+        data.get<int>("defense"),
         data.get<int>("experience_per_level"),
         data.get<int>("health_point_bonus_per_level"),
         data.get<int>("damage_bonus_per_level"),
         data.get<double>("cooldown_multiplier_per_level"),
-        data.get<double>("defense_bonus_per_level"));
+        data.get<int>("defense_bonus_per_level"));
 }
