+++ Monster.h
@@ -18,7 +18,7 @@ class Monster
 {
 
 public:
-    Monster(const std::string, const int, const int, const double, const double); ///< Constructor of Monster class
+    Monster(const std::string, const int, const int, const double, const int); ///< Constructor of Monster class
 
     /**
      * \brief Checks if the Monster's health value is greater than zero
@@ -51,14 +51,14 @@ public:
     /// This is a simple getter function for current time until next attack
     double getAttackCoolDown() const;
     /// This is a simple getter function for current defense value
-    double getDefense() const;
+    int getDefense() const;
 
     protected:
     const std::string name; ///< The name of a unit
     int health; ///< The health value of a unit
     int damage; ///< The damage value of a unit
     double atkcooldown; ///< The time between the unit's attacks
-    double defense; ///< The defense value of a unit
+    int defense; ///< The defense value of a unit
 
     private:
     static void RemoveChar(std::string & /** [in] The target string the charater(s) need to be removed from */, const char /** [in] The charater to be removed*/); ///< Removes specified characters from a given string. Used by the parseMonster function
